requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
openai==1.54.0
streamlit==1.39.0
requests==2.32.3
python-multipart==0.0.12
pydantic==2.9.0
python-dotenv==1.0.1
pytest==8.3.3
httpx==0.27.2
prometheus-client==0.21.0
google-generativeai==0.8.3
Pillow==10.4.0
